+++ src/config/trainerConfig.js
@@ -1343,7 +1343,8 @@ const expMultiplier = (level) => {
     return 3 * (Math.pow(level, 1/2));
 }
 
-const NUM_DAILY_REWARDS = process.env.STAGE === stageNames.ALPHA ? 100 : 3;
+// TEMP: gen 2 event
+const NUM_DAILY_REWARDS = (process.env.STAGE === stageNames.ALPHA ? 100 : 3) * 2;
 const NUM_DAILY_SHARDS = process.env.STAGE === stageNames.ALPHA ? 100 : 5;
 
 module.exports = {